+++ src/components/CreateChallenge.js
@@ -1,7 +1,7 @@
 import React,{Component} from 'react'
 // import propTypes from 'prop-types'
 import { graphql, gql, compose } from 'react-apollo'
-import feedQuery from './queries/feedQuery'
+import feedQuery from '../queries/feedQuery'
 
 class CreateChallenge extends Component{
