+++ src/components/Posts.js
@@ -1,7 +1,7 @@
 import React,{Component} from 'react'
 import propTypes from 'prop-types'
 import { graphql, gql, compose } from 'react-apollo'
-import feedQuery from './queries/feedQuery'
+import feedQuery from '../queries/feedQuery'
 // import gql from 'graphql-tag'
 import Paper from 'material-ui/Paper'
 import IconButton from 'material-ui/IconButton'
